+++ app/src/main/java/com/example/maniaksearch/callapi/ApiLinearAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.maniaksearch.overview
+package com.example.maniaksearch.callapi
 
 import android.util.Log
 import android.view.LayoutInflater
