+++ app/src/main/java/com/example/maniaksearch/callapi/ItunesResFragment.kt
@@ -1,16 +1,16 @@
-package com.example.maniaksearch.overview
+package com.example.maniaksearch.callapi
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.viewModels
-import com.example.maniaksearch.databinding.FragmentOverviewBinding
+import com.example.maniaksearch.databinding.FragmentItunesResBinding
 
 /**
- * Fragment showing the status of the Itunes api call to let the user know if it has failed.
+ * Fragment showing the result of the Itunes api call
  */
-class OverviewFragment() : androidx.fragment.app.Fragment() {
+class ItunesResFragment() : androidx.fragment.app.Fragment() {
 
     /**
      * Static factory creating the fragment to avoid having a constructor with arguments
@@ -20,20 +20,22 @@ class OverviewFragment() : androidx.fragment.app.Fragment() {
         const val QUERY = "query"
         const val QUERY_PARAM = "query_param"
 
-        fun newInstance(query : String, queryParam: HashMap<String, String>): OverviewFragment {
+        fun newInstance(query: String, queryParam: HashMap<String, String>): ItunesResFragment {
             val args = Bundle()
             args.putString(QUERY, query)
             args.putSerializable(QUERY_PARAM, queryParam)
-            val fragment = OverviewFragment()
+            val fragment = ItunesResFragment()
             fragment.arguments = args
             return fragment
         }
     }
 
-    private val viewModel: OverviewViewModel by viewModels()
-    var  queryParam: HashMap<String, String> = HashMap()
+    private val viewModel: ItunesResViewModel by viewModels()
+    var queryParam: HashMap<String, String> = HashMap()
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        // Get query data from the main activity by the arguments attribute
         arguments?.let {
             viewModel.query.value = it.getString(QUERY).toString()
             queryParam = it.getSerializable(QUERY_PARAM) as HashMap<String, String>
@@ -42,21 +44,20 @@ class OverviewFragment() : androidx.fragment.app.Fragment() {
 
 
     override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
+            inflater: LayoutInflater, container: ViewGroup?,
+            savedInstanceState: Bundle?
     ): View? {
-        val binding = FragmentOverviewBinding.inflate(inflater)
+        val binding = FragmentItunesResBinding.inflate(inflater)
         binding.lifecycleOwner = this
         binding.viewModel = viewModel
         binding.apiResRecyclerView.adapter = ApiLinearAdapter()
-        viewModel.query.observe(viewLifecycleOwner, { newQuery ->
-           viewModel.getApiResults(newQuery, queryParam)
-        }
+        viewModel.query.observe(
+                viewLifecycleOwner,
+                { newQuery -> viewModel.getApiResults(newQuery, queryParam) }
         )
 
         return binding.root
     }
-
 
 
 }
