+++ app/src/main/res/layout/fragment_itunes_res.xml
@@ -4,43 +4,44 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <data>
+
         <variable
             name="viewModel"
-            type="com.example.maniaksearch.overview.OverviewViewModel" />
+            type="com.example.maniaksearch.callapi.ItunesResViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="com.example.maniaksearch.overview.OverViewFragment">
+        tools:context="com.example.maniaksearch.callapi.ItunesResFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/api_res_recycler_view"
-            android:padding="6dp"
             android:layout_width="0dp"
             android:layout_height="0dp"
-            android:scrollbars="vertical"
             android:clipToPadding="false"
+            android:padding="6dp"
+            android:scrollbars="vertical"
             app:layoutManager="LinearLayoutManager"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
             app:layout_constraintTop_toTopOf="parent"
             app:listApiRes="@{viewModel.res.results}"
+            tools:itemCount="20"
             tools:layout_editor_absoluteX="183dp"
             tools:layout_editor_absoluteY="247dp"
-            tools:itemCount="20"
-            tools:listitem="@layout/itunes_res_item"/>
+            tools:listitem="@layout/itunes_res_item" />
 
         <ImageView
             android:id="@+id/status_image"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            app:apiCallStatus="@{viewModel.status}"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:apiCallStatus="@{viewModel.status}" />
+            app:layout_constraintTop_toTopOf="parent" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
