+++ app/src/main/res/menu/menu_layout.xml
@@ -2,6 +2,13 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
+    <item android:id="@+id/menu_filters"
+        android:title="@string/filters"
+        android:icon="@drawable/ic_filter"
+        android:enabled="true"
+        app:showAsAction="always">
+    </item>
+
     <item android:id="@+id/menu_param"
         android:title="@string/param"
         android:icon="@drawable/ic_param"
